backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pydantic>=2.0.0
httpx>=0.25.0
openai>=1.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
uvicorn[standard]>=0.25.0
asgiref>=3.7.0
requests>=2.31.0
tenacity>=8.0.0
